+++ src/app/dataset-view/additional-components/metadata-component/components/set-transform/components/engine-section/engine-section.component.ts
@@ -7,8 +7,8 @@ import {
     OnInit,
     Output,
 } from "@angular/core";
-import { EngineDesc, EnginesQuery, Maybe, TransformSql } from "src/app/api/kamu.graphql.interface";
-import { MaybeNull } from "src/app/common/app.types";
+import { EngineDesc, EnginesQuery, TransformSql } from "src/app/api/kamu.graphql.interface";
+import { MaybeNull, MaybeNullOrUndefined } from "src/app/common/app.types";
 import { BaseComponent } from "src/app/common/base.component";
 import { EngineService } from "src/app/services/engine.service";
 
@@ -20,7 +20,7 @@ import { EngineService } from "src/app/services/engine.service";
 })
 export class EngineSectionComponent extends BaseComponent implements OnInit {
     @Input() public knownEngines: MaybeNull<EngineDesc[]>;
-    @Input() public currentSetTransformEvent: Maybe<TransformSql> | undefined;
+    @Input() public currentSetTransformEvent: MaybeNullOrUndefined<TransformSql>;
     @Input() public selectedEngine: string;
     @Output() public onEmitSelectedEngine: EventEmitter<string> = new EventEmitter<string>();
     public selectedImage: string;